feat: Initial commit
Initial commit to get events directly from the browser. Note that n order for this to wrk we need to do some changes to the API.

QA: It has unit test and end to end tests. It was tested on Desktop on Chrome, Firefox and Safari and in Androidon Chrome, Brave, Firefox, Edge, Samsung, Opera.

In the future that test coudl be automated (partially) by using playwright.