feat: userId api
Now we allow marketplaces to set, reset and retrieve the user id. This will be useful for marketplaces to set user ids according to their own rules and also to fetch the userId for browser generated XHR requests.

Note that as we are storing the user id as a cookie, calls to theiw website will have the tsuid value set.