feat: split sku and auction attribute
We had only one attribute sku which was encoding both the sku and the acution, but that could be harder to understand for marketplaces. The reason this was donde initilaly like taht was to avoid being blocked. However, it was pointed that we stil could be blcoked, so it'd be better to remove complexity.

We also rename the attribute sku to product, as that's the name we use throughtout our documentation.