feat: split sku and auction attribute (#20)
We had only one attribute sku which was encoding both the sku and the auction, but that could be harder to understand for marketplaces. The reason this was donde initially like that was to avoid being blocked. However, it was pointed that we still could be blocked, so it'd be better to remove complexity.

We also rename the attribute sku to product, as that's the name we use throughout our documentation.

Notes:
- I'm not convinced whether we should use `data-ts-product-id` or `data-ts-product`. I started with the former, but then found a little verbose to include the `-id` suffix for two attributes.
- Decided against having these attributes names configurable. The reason is because now we added a `ts-` prefix and also in the case of BabyTuto we still need to do some changes to the markup.

# Why we need to change BabyTuto's markup?

The way the library works is by attaching a click listener to the element with the expected data attributes.

In the case of Babytuto, the `data-topsort-sku` is on a div which includes the buttons at the very bottom. Those buttons are not an action on the respective item. So they should be outside of the markup.
 
<img width="238" alt="Screen Shot 2022-02-10 at 8 16 05 PM" src="https://user-images.githubusercontent.com/911768/153512927-b33874a3-b294-4e44-8c09-eff473ac9678.png">

ps: note that currently the implemented solution only triggers a click when clicking on the image. It does not trigger a click if a user clicks on either the Brand, the product name or the price.

ps2: one limitation of the analytics.js library is that the whole element is considered clickable, which is not always the case. For example in babytuto all the whitespace and even the discount tag are not clickable.