chore: bump dev deps
We manually bump the dev dependencies, except for jest, as it currently does not work with ts-jest.

We also change the dependabot config, so that dev dependencies are also updated in the package.json file. This is required as now the package yis a public library and otherwise dependabot would only update the package-lock.json file. See https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#versioning-strategy