feat: propagate additional products (#213)
In case customers want to attribute purchases to banners they need to
specify a product and a resolved bid as inherit. In such a case we would
grab the resolved id id that was stored in session storage in the
previous click.

There are some caveats with this approach:
1) If there's a mistake and an inherit resolved bid is used in a page
which is not the target of a banner, then the events will fail. 2) It
could happen that the user manually changes the url to a banner's
destination page and in that case we would report the events to the
wrong banner. 3) This won't always work if session storage is not
available (support is 97.4%
https://caniuse.com/mdn-api_window_sessionstorage). We have a fallback
using an in memory cache, but that would only work if the page it's not
reloaded (Vue, React, etc)